major changes for v0.4.1
