+++ client/stylesheets/classes.scss
@@ -1,3 +1,10 @@
+.centered {
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  margin-top: -50px;
+  margin-left: -100px;
+}
 .display-flex {
   display: flex;
 }
@@ -30,6 +37,10 @@
 .img-square-35 {
   width: 35px;
   height: 35px;
+}
+.img-square-40 {
+  width: 40px;
+  height: 40px;
 }
 .margin-bottom-5 {
   margin-bottom: 5px !important;
