+++ client/stylesheets/scrollbars.scss
@@ -1,0 +1,37 @@
+.scrollbar-dark::-webkit-scrollbar-track
+ {
+   -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
+   background-color: #9d99a0;
+ }
+
+.scrollbar-dark::-webkit-scrollbar
+{
+  width: 12px;
+  background-color: #9d99a0;
+}
+
+.scrollbar-dark::-webkit-scrollbar-thumb
+{
+  border-radius: 10px;
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,.3);
+  background-color: #555;
+}
+
+.scrollbar-light::-webkit-scrollbar-track
+{
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,0.3);
+  background-color: #f1f1f1;
+}
+
+.scrollbar-light::-webkit-scrollbar
+{
+  width: 12px;
+  background-color: #f1f1f1;
+}
+
+.scrollbar-light::-webkit-scrollbar-thumb
+{
+  border-radius: 10px;
+  -webkit-box-shadow: inset 0 0 6px rgba(0,0,0,.3);
+  background-color: #C1C1C1;
+}
