--- client/templates/app/loading.html
@@ -1,3 +1,0 @@
-<template name="loading">
-	
-</template>
