+++ client/templates/chat/_sidebar/settings.html
@@ -23,16 +23,6 @@
       </div>
 
       <div style="padding-bottom: 5px;">
-        <div class="display-inline-block white-text" style="width: 196px;">My Messages Right</div>
-        <div class="display-inline-block switch">
-          <label>
-            <input class="switch-setting" type="checkbox" data-name="ownerMessagesLeft" disabled/>
-            <span class="lever no-margin"></span>
-          </label>
-        </div>
-      </div>
-
-      <div style="padding-bottom: 5px;">
         <div class="display-inline-block white-text" style="width: 196px;">Turn Off Chat Colors</div>
         <div class="display-inline-block switch">
           <label>
