+++ client/templates/chat/home.html
@@ -8,10 +8,6 @@
           </div>
           <div class="flex-1-1-auto fill-vh fill-width pos-rel">
             {{> chatMainPanelWindow}}
-            {{#if Template.subscriptionsReady}}
-            {{else}}
-              {{> loading}}
-            {{/if}}
           </div>
 
       </div>
