+++ imports/template_helpers/_helpers/markdown.js
@@ -1,0 +1,7 @@
+import { Meteor } from 'meteor/meteor';
+import { Template } from 'meteor/templating';
+
+import markdown from 'markdown';
+
+Template.registerHelper('markdown',
+  (text) => (text && text.length) ? markdown.parse(text) : "");
