+++ imports/template_helpers/admin/index.js
@@ -1,0 +1,1 @@
+import './admin.js';
