+++ imports/template_helpers/app/index.js
@@ -1,0 +1,1 @@
+import './app.js';
