+++ imports/template_helpers/chat/index.js
@@ -1,0 +1,3 @@
+import './onload.js';
+import './helpers.js';
+import './events.js';
