+++ imports/template_helpers/home/index.js
@@ -1,0 +1,1 @@
+import './home.js';
