+++ public/docs/COMMANDS.md
@@ -1,0 +1,7 @@
+## Chat Commands
+
+* /help
+* /image
+* /purge
+* /time (ALT + T)
+* /topic
