+++ public/docs/ISSUES.md
@@ -1,8 +1,9 @@
-# :boom: Meteor Blast :boom:
-
 ## Known Issues
 
 * __Major__: Mobile nav login and other functionality broken
+* Admin drop down menu not always initializing
 * Scrolling all the way to bottom on load
 * Chat layout broken when adding "nav" back in
 * Much more latency using imports for templates?
+* ALT+ Commands across browsers (IE Support)
+* Add username support for Oauth services
